+++ src/styles/fonts.js
@@ -1,2 +1,3 @@
 import "typeface-bungee-shade"
 import "typeface-montserrat"
+import "typeface-courier-prime"
